fixes #21 by creating/deleting the directory in the provider rather than the factory.
