+++ pom.xml
@@ -31,8 +31,8 @@
     <main.java.package>io.phasetwo.keycloak.themes</main.java.package>
     <freemarker.version>2.3.31</freemarker.version>
     <junit.version>5.8.2</junit.version>
-    <keycloak.version>24.0.0</keycloak.version>
-    <lombok.version>1.18.30</lombok.version>
+    <keycloak.version>24.0.3</keycloak.version>
+    <lombok.version>1.18.32</lombok.version>
     <auto-service.version>1.1.1</auto-service.version>
     <ossrh.url>https://s01.oss.sonatype.org</ossrh.url>
   </properties>
